pyproject.toml
[project]
name = "purple-agent-wirelessagent"
version = "0.1.0"
description = "Purple Agent for UC Berkeley AgentX Competition - Wireless Communication Expert"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "uvicorn>=0.38.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
